requirements.txt
django
djangorestframework
pygments
drf-spectacular